+++ base/applications/rapps/lang/fr-FR.rc
@@ -210,6 +210,9 @@ BEGIN
     IDS_SELECTEDFORINST "Sélectionnées pour installation"
     IDS_MISMATCH_CERT_INFO "Certificat inconnu:\nSujet: %s\nEmetteur: %s\nVoulez-vous continuer ?"
     IDS_UNABLE_PATH "Format de chemin invalide."
+    IDS_APP_DISPLAY_DETAILS "details"
+    IDS_APP_DISPLAY_LIST "list"
+    IDS_APP_DISPLAY_TILE "tile"
 END
 
 STRINGTABLE
