+++ base/applications/rapps/lang/ja-JP.rc
@@ -210,6 +210,9 @@ BEGIN
     IDS_SELECTEDFORINST "インストールを選択済み"
     IDS_MISMATCH_CERT_INFO "証明書は未知のものです:\nSubject: %s\nIssuer: %s\nそれでも続けますか？"
     IDS_UNABLE_PATH "Incorrect path format."
+    IDS_APP_DISPLAY_DETAILS "details"
+    IDS_APP_DISPLAY_LIST "list"
+    IDS_APP_DISPLAY_TILE "tile"
 END
 
 STRINGTABLE
