+++ base/applications/rapps/lang/zh-TW.rc
@@ -212,6 +212,9 @@ BEGIN
     IDS_SELECTEDFORINST "選中以安裝"
     IDS_MISMATCH_CERT_INFO "所使用的證書未知：\n主題：%s\n發行方：%s\n您是否仍然想繼續？"
     IDS_UNABLE_PATH "不正確的URL格式。"
+    IDS_APP_DISPLAY_DETAILS "details"
+    IDS_APP_DISPLAY_LIST "list"
+    IDS_APP_DISPLAY_TILE "tile"
 END
 
 STRINGTABLE
